updating

+++ zoodatasets/chunkdatasets.py
@@ -121,7 +121,7 @@ class ZooDataset(Dataset):
                         wl.append(w)
                 else:
                     wl.append(w)
-            data = torch.cat(wl, dim=0)
+            data = torch.cat(wl, dim=0).to(torch.bfloat16)
         else:
             w = data[dataset]
             w = pad_to_chunk_multiple(w, chunk_size=self.chunk_size)
